+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Version 1.1.0
 * adds Ribbon integration
+* exponential backoff customizable via Retryer.Default ctor
 
 ### Version 1.0.0